New classes RefactoringClassEvents and RefactoringMethodEvents extracted from class RefactoringEvents.
